# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Device Management & Measurement API",
    "description": "API to manage customers, devices, and ingest/retrieve measurements.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/devices": {
      "post": {
        "tags": [
          "Devices"
        ],
        "summary": "Create a new device",
        "description": "Register a new device. The device is created without a customer assignment.\n\n- `serial_number` must be unique across all devices.\n- Returns **409** if the serial number is already registered.\n- Returns **422** if payload validation fails.",
        "operationId": "create_device_devices_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                },
                "example": {
                  "detail": "A resource with the same unique value already exists"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                },
                "example": {
                  "detail": [
                    {
                      "loc": [
                        "body",
                        "email"
                      ],
                      "msg": "value is not a valid email address: The email address is not valid. It must have exactly one @-sign.",
                      "type": "value_error",
                      "input": "invalid-email"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/devices/{device_id}/customer": {
      "put": {
        "tags": [
          "Devices"
        ],
        "summary": "Assign device to a customer",
        "description": "Assign a device to a customer, optionally setting a timezone.\n\n- Idempotent: re-assigning to the same or a different customer is always accepted.\n- Returns **404** if the device or customer does not exist.\n- Returns **422** if payload validation fails.",
        "operationId": "assign_device_customer_devices__device_id__customer_put",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Device Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignCustomerRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Resource with the given id was not found"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "example": {
                  "detail": [
                    {
                      "loc": [
                        "body",
                        "email"
                      ],
                      "msg": "value is not a valid email address: The email address is not valid. It must have exactly one @-sign.",
                      "type": "value_error",
                      "input": "invalid-email"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Devices"
        ],
        "summary": "Unassign device from its customer",
        "description": "Remove the customer assignment from a device and clear its timezone.\n\n- Idempotent: unassigning an already-unassigned device is a successful no-op.\n- Returns **404** if the device does not exist.",
        "operationId": "unassign_device_customer_devices__device_id__customer_delete",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Device Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Resource with the given id was not found"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/devices/{device_id}/activation": {
      "patch": {
        "tags": [
          "Devices"
        ],
        "summary": "Update device activation status",
        "description": "Update the activation status of a device (activate or deactivate).\n\n- Setting `is_active: true` enables measurement ingestion.\n- Setting `is_active: false` prevents further measurement ingestion.\n- Idempotent: repeating the same target state is a successful no-op.\n- Returns **404** if the device does not exist.\n- Returns **422** if payload validation fails.",
        "operationId": "update_device_activation_devices__device_id__activation_patch",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Device Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceActivationUpdate"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Resource with the given id was not found"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "example": {
                  "detail": [
                    {
                      "loc": [
                        "body",
                        "email"
                      ],
                      "msg": "value is not a valid email address: The email address is not valid. It must have exactly one @-sign.",
                      "type": "value_error",
                      "input": "invalid-email"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/devices/{device_id}/measurements": {
      "post": {
        "tags": [
          "Devices"
        ],
        "summary": "Ingest device measurements",
        "description": "Ingest a batch of measurements for a device.\n\n- The client provides a unique `measurement_id` per measurement for idempotency.\n- Duplicate `measurement_id` values are silently ignored (no duplicates created).\n- Accepts between 1 and 1,000 measurements per request. Larger payloads must be chunked.\n- Returns **201 Created** on success (even for fully duplicate batches).\n- Returns **404** if the device does not exist.\n- Returns **422** if the device is inactive or payload validation fails.",
        "operationId": "ingest_measurements_devices__device_id__measurements_post",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Device Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/MeasurementIngestItem"
                },
                "minItems": 1,
                "maxItems": 1000,
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Resource with the given id was not found"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "example": {
                  "detail": [
                    {
                      "loc": [
                        "body",
                        "email"
                      ],
                      "msg": "value is not a valid email address: The email address is not valid. It must have exactly one @-sign.",
                      "type": "value_error",
                      "input": "invalid-email"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Devices"
        ],
        "summary": "List device measurements",
        "description": "List measurements for a device with optional filters and cursor-based pagination.\n\n- Results are sorted by timestamp, most recent first by default.\n- Filters combine with AND logic.\n- Returns **404** if the device does not exist.\n- Returns **422** if query parameters fail validation.",
        "operationId": "list_measurements_devices__device_id__measurements_get",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Device Id"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by exact measurement type",
              "title": "Type"
            },
            "description": "Filter by exact measurement type"
          },
          {
            "name": "start_time",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Inclusive lower bound on timestamp (UTC)",
              "title": "Start Time"
            },
            "description": "Inclusive lower bound on timestamp (UTC)"
          },
          {
            "name": "end_time",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Exclusive upper bound on timestamp (UTC)",
              "title": "End Time"
            },
            "description": "Exclusive upper bound on timestamp (UTC)"
          },
          {
            "name": "sort_dir",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(asc|desc)$",
              "description": "Sort direction",
              "default": "desc",
              "title": "Sort Dir"
            },
            "description": "Sort direction"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Maximum results per page",
              "default": 20,
              "title": "Limit"
            },
            "description": "Maximum results per page"
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque pagination cursor from previous response",
              "title": "Cursor"
            },
            "description": "Opaque pagination cursor from previous response"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedMeasurementsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Resource with the given id was not found"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "example": {
                  "detail": [
                    {
                      "loc": [
                        "body",
                        "email"
                      ],
                      "msg": "value is not a valid email address: The email address is not valid. It must have exactly one @-sign.",
                      "type": "value_error",
                      "input": "invalid-email"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeviceActivationUpdate": {
        "properties": {
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "Target activation status of the device"
          }
        },
        "type": "object",
        "required": [
          "is_active"
        ],
        "title": "DeviceActivationUpdate",
        "description": "Schema for updating a device's activation status."
      },
      "DeviceCreate": {
        "properties": {
          "serial_number": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Serial Number",
            "description": "Unique device serial number"
          }
        },
        "type": "object",
        "required": [
          "serial_number"
        ],
        "title": "DeviceCreate",
        "description": "Schema for creating a new device."
      },
      "DeviceResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Unique device identifier"
          },
          "serial_number": {
            "type": "string",
            "title": "Serial Number",
            "description": "Unique device serial number"
          },
          "customer_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Id",
            "description": "ID of the assigned customer; null if unassigned"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Device lifecycle status (e.g. 'unassigned', 'assigned')"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "Whether the device is active and can ingest measurements"
          },
          "timezone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone",
            "description": "IANA Time Zone identifier assigned to this device"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "UTC timestamp when the device was registered"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "UTC timestamp of the last update"
          }
        },
        "type": "object",
        "required": [
          "id",
          "serial_number",
          "customer_id",
          "status",
          "is_active",
          "created_at",
          "updated_at"
        ],
        "title": "DeviceResponse",
        "description": "Schema for device response."
      },
      "ErrorDetail": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Human-readable description of the error"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorDetail",
        "description": "Body returned by all domain-level error handlers (404, 409, 500)."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MeasurementIngestItem": {
        "properties": {
          "measurement_id": {
            "type": "string",
            "format": "uuid",
            "title": "Measurement Id",
            "description": "Client-generated unique ID for idempotency"
          },
          "type": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Type",
            "description": "Measurement type (e.g. 'temperature')"
          },
          "value": {
            "type": "number",
            "title": "Value",
            "description": "Numeric measurement value"
          },
          "unit": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Unit",
            "description": "Unit of measurement (e.g. '\u00b0C')"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp",
            "description": "UTC timestamp of the measurement"
          }
        },
        "type": "object",
        "required": [
          "measurement_id",
          "type",
          "value",
          "unit",
          "timestamp"
        ],
        "title": "MeasurementIngestItem",
        "description": "Schema for a single measurement within an ingestion request."
      },
      "MeasurementResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Unique measurement identifier"
          },
          "device_id": {
            "type": "string",
            "format": "uuid",
            "title": "Device Id",
            "description": "ID of the device that produced this measurement"
          },
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Measurement type (e.g. 'temperature', 'energy')"
          },
          "value": {
            "type": "number",
            "title": "Value",
            "description": "Numeric measurement value"
          },
          "unit": {
            "type": "string",
            "title": "Unit",
            "description": "Unit of measurement (e.g. '\u00b0C', 'kWh')"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp",
            "description": "UTC timestamp when the measurement was recorded"
          }
        },
        "type": "object",
        "required": [
          "id",
          "device_id",
          "type",
          "value",
          "unit",
          "timestamp"
        ],
        "title": "MeasurementResponse",
        "description": "Schema for a single measurement in a list response."
      },
...
      "PaginatedMeasurementsResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MeasurementResponse"
            },
            "type": "array",
            "title": "Items",
            "description": "Measurements on this page"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor",
            "description": "Opaque cursor to pass as `cursor` to retrieve the next page; null if no more pages"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "description": "Whether additional pages exist"
          }
        },
        "type": "object",
        "required": [
          "items",
          "next_cursor",
          "has_more"
        ],
        "title": "PaginatedMeasurementsResponse",
        "description": "Paginated list of measurements."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ValidationErrorItem": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Loc",
            "description": "Path to the failing field, e.g. ['body', 'email']"
          },
          "msg": {
            "type": "string",
            "title": "Msg",
            "description": "Human-readable reason the validation failed"
          },
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Machine-readable error type identifier"
          },
          "input": {
            "title": "Input",
            "description": "The value that was rejected"
          },
          "ctx": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ctx",
            "description": "Extra context from the validator"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationErrorItem",
        "description": "A single Pydantic / FastAPI validation failure."
      },
      "ValidationErrorResponse": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationErrorItem"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ValidationErrorResponse",
        "description": "Body returned on 422 Unprocessable Content."
      }
    }
  }
}